Add files via upload
The barycenter indices of a heatmap intensity information (C, in Matlab imagesc) can be calculated using the following methods:
(1) https://www.mathworks.com/matlabcentral/answers/133395-compute-centroid-of-a-matrix
(2) https://www.mathworks.com/matlabcentral/answers/363181-center-of-mass-and-total-mass-of-a-matrix
Then perform a linear fit between the original coordinate vectors x, y and the index vectors xIdx, yIdx to get the slopes and intercepts for both fits (Note: you could also use the function p = polyval(x,y,1) to calculate slope (p(1)) and intercept (p(2))). Use the calculated indices for barycenterIdxX and barycenterIdxY of either (1) or (2) to evaluate the linear fit to get the barycenter values (barycenterX, barycenterY; assuming that x and y are evenly spaced).